chore: move URLs to their own separate files
- This should help clean up the code of the components a bit more
