+++ src/lib/Hukamnama.svelte
@@ -1,16 +1,13 @@
 <script lang="ts">
   import isEmpty from 'lodash/isEmpty';
+  import { getHukamnamaUrl } from '../urls';
 
   export let date = new Date();
 
   let hukamnamaPromise = getHukamnama();
 
   async function getHukamnama() {
-    const day = String(date.getDate()).padStart(2, '0');
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const year = date.getFullYear();
-
-    return fetch(`/raw_hukamnamas/${year}${month}${day}.html`, { method: 'GET' })
+    return fetch(getHukamnamaUrl(date), { method: 'GET' })
       .then(res => res.text())
       .then(text => new DOMParser().parseFromString(text, 'text/html'))
       .then(res => { console.log({ textAsHTML: res }); return res; })
